+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-10-05
+
+### Changed
+
+- Genesis code display now has improved and uniform look for light mode users.
+
 ## 2024-10-04
 
 ### Changed